Add a spec file to the imagefactory top-level.
Unfortunately, using the "bdist_rpm" target for setup.py
isn't really expressive enough to generate all of the things
we might need to do with the spec file for imagefactory.
As one example, we might need to add additional dependencies
to the spec beyond the ones that bdist_rpm detected.  It would
also be nice to have a package description and a changelog.

To fix all of that, maintain our own spec at the top of
the repository hierarchy.  That way we have full control
over it.

Signed-off-by: Chris Lalancette <clalance@redhat.com>
Signed-off-by: Steve Loranz <sloranz@redhat.com>