Add imgfac cmdline support for building and pushing multiple images
Add support to imgfac for building images for multiple target and
pushing images to multiple providers.

New 'image' and 'build' IWHD objects are introduced as outlined here:

  https://fedorahosted.org/pipermail/aeolus-devel/2011-May/001667.html

The build cmdline now looks like:

  $> imgfac --template <template> --target <target> [--target <target> ...]

The push cmdline is now:

  $> imgfac --image <image-UUID> --credentials <credentials-XML> --provider <provider> [--provider <provider> ...]

One item that requires further thought is what to do about determining
the target image appropriate for a given provider. Currently, we have a
fairly lame heuristic, but perhaps we should allow conductor to supply
this.

Still TODO in BuildDispatcher.py:

  - Split the watcher code out into separate modules
  - Make use of a singleton instead of class methods