Rename rhev-m/vmware target names to rhevm/vsphere
In the case of RHEV-M and VMWare, image factory target names differ
from deltacloud driver names. As of 3b20568b, we attempt to map
deltacloud driver names in order to support both sets of names.

However, this doesn't account for the target name which is recorded
in IWHD. If a client requests an image to be built for 'rhevm', the
target recorded in IWHD is 'rhev-m'.

We should take the opportunity and cut out all confusion by
standardising on deltacloud driver names.

We could add some compatibility code to gracefully handle the previous
driver names, but it probably isn't warranted at this point.

https://fedorahosted.org/pipermail/aeolus-devel/2011-June/002120.html