Remove a dangerous default dictionary.
As pointed out by pylint, start_builder_thread was using a
dangerous default dictionary.  Default dictionaries are
dangerous because they are initialized at the time that the
method is parsed, not at the time that it is invoked.  This
means that if you call the method twice and manipulate the
dictionary within the method, it will manipulate the first
version of the dictionary, not a brand-new one.  Luckily all
callers of the method passed the argument, so just remove the
default empty dictionary.

Signed-off-by: Chris Lalancette <clalance@redhat.com>