Split FedoraBuilder up into several parts.
Every OS,backend tuple that imagefactory supports needs
to do very different things in order to get the image
prepared for the cloud.  Because of this, it doesn't make
a lot of sense to have one large FedoraBuilder file that
contains all of the logic.  Not only is it untidy, it makes
it difficult to reason about the code paths individual OS
and cloud combinations will take.

This patch splits the massive class into <OS>_<target>_builder.py
files and classes.  Then the top-level BuildJob is adjusted
to import the right class when it needs this.  It makes
for a more readable and maintainable source tree.

Signed-off-by: Chris Lalancette <clalance@redhat.com>