more detailed response for builds
The response now includes a list of dictionaries for the builders that were started.  This includes the id, target, and a url to get the builder details from the rest api.  To make things easier to reference, everything is in one module.  This can get split again later if needed.

Signed-off-by: Steve Loranz <sloranz@redhat.com>