Added property for builder type and set bottle reloader to False.
In order to get an idea of what a builder is doing (i.e. building or pushing) I added a property to hold that info.
Also, I was seeing paste httpserver worker threads get stopped and restarted, clearing out any state.  This is because I had reloader set to True in the run() call.  Set this to False.

Signed-off-by: Steve Loranz <sloranz@redhat.com>