remove TDL name from internal naming used when building via Oz
libvirt, via Oz, uses the name field in our TDL to construct paths.  One particular
path is used as the name of Unix domain socket that monitors a qemu instance.
This path is limited to 108 characters, a limit that can easily be met by using
a long-ish name in the TDL.

There's no particular need for Oz to use a name internally that matches the
human readable name given in the TDL.  We already have to twiddle the internal
Oz name to ensure uniqueness when doing parallel builds.  This patch simply
removes any reference to the TDL name entirely.

Note that we preserve the original TDL XML when storing the template in the
warehouse, so none of this information is actually lost.