add ephemeral mappings to EBS registration to match default S3 values
When registiring an S3 image we get a default mapping of the EC2 ephemeral devices
to device names.  When registiring an EBS image, we do not.  This patch creates
mappings for the EBS registration that match the default S3 values.  This allows
us to use a single fstab per architecture for both AMI types.  We likely want to
look at making this more flexible and possibly user-defineable in the future.

Details of how to get boto to do this mapping can be found in this thread:

http://groups.google.com/group/boto-users/browse_thread/thread/ee281f1ccf420a19/26a72f3c9341526b