fix handling of scenario where rc.local does not exist
My earlier patch to address did not add a bash shebang at the start
of the new rc.local.  As a result, it fails.  This fixes that.

TODO: This rc.local stuff is getting pretty messy.  Upstream now
has a more general purpose tool to address key injection on EC2
and elsewhere.  Let's look at it:

http://lists.fedoraproject.org/pipermail/package-announce/2011-October/067146.html