allow templates with version strings
We do a simple check in Template.py to see if a template string looks
like valid XML.  That check fails if a version property is added to
the top level <template> tag.  The version string is valid and is
something we were asked to add for consistency checking.

Allow it.
(cherry picked from commit e30c802967d63e4871dabb708bb2b2d0756c1299)