allow templates with version strings
We do a simple check in Template.py to see if a template string looks
like valid XML.  That check fails if a version property is added to
the top level <template> tag.  The version string is valid and is
something we were asked to add for consistency checking.

Allow it.