fix perms issue for dc-rhev-image
The way we currently push to RHEV-M requires that the process doing the
pushing run as the kvm user and group.  That is hard coded to 36:36 in
dc-rhev-image.  Our umask change made the rhev-m image files unreadable
to the tool which makes pushes fail.

As I say in the patch, we need to figure out a cleaner way to do this.
This should be a TODO when we move the push code into factory.