Swapped out status delegate for observer pattern.
The new builder, using os and cloud plugins, does much less work on it's own if any at all.  It made less sense to call out to a delegate for status changes being set by the plugins. As a result, this was swapped out for an observer implementation using NotificationCenter and Notification.

Signed-off-by: Steve Loranz <sloranz@redhat.com>