modify plugins to use shared utility functions for common libguestfs activities
At the moment this is mostly related to removing MAC address details from the images
and tagging images with a cloud-info file.

Note that the new shared utility functions are more flexible about the disk layout
of the incoming image.  This should help when we add support for arbitrary base
image imports.

Also note that the MAC removal and cloud-info tagging were actually completely
absent from the OpenStack plugin.  I have added them.  This was a bug.