re-enable and improve ICICLE generation and propagation
At some point we stopped persisting ICICLE when building base images

This patch turns that back on and also ensures that ICICLE is carried forward
to child objects.  Note that it can still be overridden by the plugins and that
this will in fact happen if further customization is done at the target_image
stage.