rework IndirectionCloud to be compatible with Oz migration to lxml
We are tightly integrated with Oz/TinMan for this plugin.  This includes
directly passing parsed XML objects into Oz functions.  The switch to lxml
in Oz broke some of these calls.  This patch fixes it.

This also changes the RPM packaging to force Oz >= 0.12.0 which is the
first Oz release version to use the lxml changes.