make TinMan interpret string and integer parameters as bool where necessary
Per the code comments, this is needed to allow the newly expanded CLI options
for parameters passing to work correctly.

What we should really do is make each plugin specify what parameters it can
accept, what the type is and what the defaults should be and then do strict
validation or casting of the incoming data.

TODO: Add that to the backlog