first cut at Rackspace upload functionality
Rackspace uses a task-based upload scheme that, as of the time of this patch,
is not part of upstream OpenStack.  To use it we have to make use of the
pyrax Python binding.

The "provider" in the case of Rackspace is expected to be one of the standard
rackspace region codes.  Known regions supported right now are:

ORD, DFW and SYD

On the Factory CLI these should be specified as @ORD, etc.