use qcow2 with backing file for more efficient utility image launching
The qcow2 format can make use of a backing file.  Unallocated blocks are
read from the backing file and changes are saved in the qcow2 image.  The
backing file can be either a raw image or another qcow2.

This patch uses this feature to avoid making a full copy of the utility
image before launching it to execute the indirection template.  This
is faster and works particularly well in our koji integration, where the
utility image is often available via a read-only NFS mount.