ensure qcow v2 compatibility when doing qemu-img convert
As of F21 and RHEL7/CentOS 7, qemu-img defaults to v3 of qcow2
when doing "qemu-img convert" commands.  The resulting qcow files
are not backwards compatible and will break if deployed on some
older systems, including RHEL6.  These changes ensure that any
qcow2 we produce inside of Factory is v2 backwards compatible.

Note that, as of this patch, the image creation code in Oz/TinMan
produces compatible qcow files.  It may need changes in future to
ensure this remains the case.