add vagrant virtualbox and libvirt-kvm support to the OVA plugin
THis changeset adds the ability to create vagrant "box" files
for both virtualbox and libvirt based vagrant installs.

As with the rest of the formats supported by the OVA plugin, you
are expected to pass the ID of an existing rhev-m or vsphere target
image as the "--id" of your ova target_image build.

For vagrant virtualbox, pass in the ID of a vsphere target image
along with a parameter KV pair of:

'rhevm_ova_format': 'vagrant-virtualbox'

For vagrant libvirt, pass in the ID of a rhevm target image along
with a parameter KV pair of:

'vsphere_ova_format': 'vagrant-libvirt'

The boxes will use SSH insecure key authentication by default. It is
up to you to generate a base image that contains the needed authorized
key and vagrant user.

The boxes are configured for rsync sync of your host working directory
into /vagrant in the guest box.  You can specify a different default
sync directory by passing it in the 'vagrant_sync_directory' parameter.
We expect to use this when generating Atomic images, where the /
filesystem is not writeable at runtime.

You can also override any of these settings in your local Vagrantfile.
See the Vagrant documentation for more information.