Add support for VMWare Fusion Vagrant boxes
This works in the same way as the VirtualBox and libvirt Vagrant support.
For VMWare fusion, pass the following via the CLI when creating the final
OVA target image:

--parameter vsphere_ova_format vagrant-vmware-fusion

Then input in this case needs to be a vSphere target image using the normal
non-streaming VMDK format.  To create this, pass the following via the CLI
during the vSphere target image creation step:

--parameter vsphere_vmdk_format standard

(Note that the default behavior for vSphere continues to be to create a
streaming VMDK file, which can be used when generating vSphere OVAs and
when pushing images directly to vSphere servers.  The parameter value
for this is "streaming".)