stop-gap change to allow RHEL7 EC2 builds
This helper class deals with the ability to modify and snapshot
remote AMIs in EC2.  However, we attempt to import the associated
class in all cases, even when we are uploading an image.  Add an
empty RHEL7 implementation to allow uploads to work for now.

TODO: Revisit remote modification and snapshots.