do not require a root PW in TDL by default
This was originally implemented for the "TDL only without kickstart"
use case.  That use case is now very uncommon.  The net result is
that users are required to set a password in the TDL that does not
actually end up getting used in the installed image. This is silly.

If you are using Factory without passing an install script (kickstart,
autoinstall, etc.) please set a root/admin password in the TDL or you
risk getting a well known default from Oz.