add tar_excludes parameter to allow pruning content from Docker images
Users may now pass in a comma separated list of excludes that will be
passed as a series of --exclude arguments to tar.  See tar man page for
details on what this does and how to use things like wildcards.

This allows images to be pruned after they have been built.  This is
particularly important if removing content that might prevent the image
from booting or even being properly inspected by libguestfs.  One such
piece of content is fstab, which is the default value for this new
parameter.