add initial support for "v2" base images
This is not really the same thing as v2 images in a remote registry.
However, this new format is a significant change and may become the
only supported single file "load" mechanism at some point in the
future.

To use this format pass a value of "1.10.1" to the "dockerversion"
parameter when generating a target image using the docker plugin.