Initial drop of HyperV Vagrant functionality
This works in the same way as our support for other Vagrant box types.
In this case the initial target is "hyperv" and the follow on target
is "ova".

Because of uncertainty about the status of rsync support in hyperv
Vagrant, this plugin currently disables sync in the default Vagrantfile
to avoid a default experience that causes an error.  It can be turned
back on with a custom Vagrantfile.