Add "ec2_image_format" configuration variable.
This allows you to use the EC2 VM import service. You would set the
"ec2_image_format" configuration variable to a supported format, and then
pass

  --parameter skip_ec2_modifications true

to the "target_image" command. The result is an image that can be used with
the EC2 image import service.

Note that in theory this change is not strictly necessary, as the VM import
service does support the RAW format. However RAW images are very large and
in our case it's not a practical format for our users to download. So a
sparse format like VMDK or VHD is preferred.

[1] http://docs.aws.amazon.com/vm-import/latest/userguide/vmimport-image-import.html