Future-import print_function, fix print calls
Apart from the future import where print is a function, not a statement, fix
incorrect print formatting calls. Calls like print("%s" % (foo)) are subtly
broken, depending on the value of foo, since (foo) is not a tuple but really
syntactic no-op. The correct way to do that is to use print("%s" % (foo,))

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>